requirements.txt
pygame
PyAutoGUI
keyboard
ollama
openai-whisper
coqui-tts
Wave
PyAudio
threading